Importers: Rename site-importer/site-importer
