+++ client/my-sites/importer/importer-site-importer.jsx
@@ -11,7 +11,7 @@ import React from 'react';
 /**
  * Internal dependencies
  */
-import SiteImporter from './site-importer/site-importer';
+import SiteImporter from './site-importer';
 
 class ImporterSiteImporter extends React.PureComponent {
 	static displayName = 'ImporterSiteImporter';
@@ -20,6 +20,8 @@ class ImporterSiteImporter extends React.PureComponent {
 		title: 'Wix.com (Beta)',
 		icon: 'site-importer',
 		description: this.props.translate( 'Import posts, pages, and media from your Wix.com site.' ),
+		// TODO: we could move this to the component itself. Here were trying to stick to a generalisation
+		// that doesn't really apply for this importer - we don't upload anything as such.
 		uploadDescription: this.props.translate( 'Type your existing site URL to start the import.' ),
 	};
